# Remove `fmt` package dependency