This pull request removes `fmt` as a dependency, reducing library size and adding some speed improvements.

First I've replaced `fmt.Sprintf("%x", var)` with `hex.EncodeToString(var)`. The [`encoding/hex`](https://golang.org/pkg/encoding/hex/) is the right fit, as _implements hexadecimal encoding and decoding_.

Then instead of calling `fmt.Errorf("...")` to create an error, I've extracted the error to a variable initialized using `errors.New`.

## Differences in library size

After `go build` and `go install`, the sizes in bytes reported by `du -b` are `12450` for current master and `8542` for this branch. That's almost a 32% size reduction.

## Performance improvements

There are more improvements on allocations and memory consumption than in speed.

I've run benchmarks for 60 seconds in both cases, with this command: `go test -bench . -benchmem -benchtime 60s`.

In current `master`:

```
PASS
BenchmarkCheck-4  	   10000	  13747472 ns/op	    1288 B/op	      15 allocs/op
BenchmarkEncrypt-4	   10000	  13792760 ns/op	    1432 B/op	      16 allocs/op
ok  	github.com/inkel/shield	278.268s
```

In branch `remove-fmt`:

```
PASS
BenchmarkCheck-4  	   10000	  13658628 ns/op	    1136 B/op	      10 allocs/op
BenchmarkEncrypt-4	   10000	  13653159 ns/op	    1280 B/op	      11 allocs/op
ok  	github.com/inkel/shield	275.968s
```

Comparing the benchmarks with `benchcmp` these are the results:

```
benchmark              old ns/op     new ns/op     delta
BenchmarkCheck-4       13747472      13658628      -0.65%
BenchmarkEncrypt-4     13792760      13653159      -1.01%

benchmark              old allocs     new allocs     delta
BenchmarkCheck-4       15             10             -33.33%
BenchmarkEncrypt-4     16             11             -31.25%

benchmark              old bytes     new bytes     delta
BenchmarkCheck-4       1288          1136          -11.80%
BenchmarkEncrypt-4     1432          1280          -10.61%
```

As you can see the speed improvements aren't much, but the differences in allocations and bytes used  are worth merging this branch.